requirements-dev.txt
isort==5.8.0
black==20.8b1
flake8==3.9.0
flake8-django==1.1.1
flake8-bugbear==21.4.3
